updated setNeutralModeFX in TalonFX.c
The earlier CAN packet we documented was actually wrong, sorry :(

Updated the CAN packet that was being sent and added logic to send the packet twice if you want to apply coast mode. We noticed during sniffing that you need to send this packet twice in order for the change to happen correctly. 

Again had to use a delay. Would be nice to learn how to work around using delays. This function wouldn't work otherwise.